reco-builder/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
sqlalchemy>=1.4.49,<2.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
numpy>=1.25.0
pandas>=2.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
minio>=7.2.7
scikit-learn>=1.3.0